+++ changelog.md
@@ -2,13 +2,14 @@
 
 ## What's Changed
 
+- Add schedule support to custom data point
 - Optimize climate get/set_schedule
 
 # Version 2025.11.15 (2025-11-13)
 
 ## What's Changed
 
-- Move reload_and_cache_schedules to load_data_point_value
+- Move reload_and_cache_schedule to load_data_point_value
 
 # Version 2025.11.14 (2025-11-13)